+++ tests/framework/validators/BooleanValidatorTest.php
@@ -8,7 +8,6 @@
 namespace yiiunit\framework\validators;
 
 use yii\validators\BooleanValidator;
-use yii\web\View;
 use yiiunit\data\validators\models\FakedValidationModel;
 use yiiunit\TestCase;
 
@@ -77,23 +76,5 @@ class BooleanValidatorTest extends TestCase
         $validator->validate('someIncorrectValue', $errorMessage);
 
         $this->assertEquals('the input value must be either "true" or "false".', $errorMessage);
-
-        $obj = new FakedValidationModel();
-        $obj->attrA = true;
-        $obj->attrB = '1';
-        $obj->attrC = '0';
-        $obj->attrD = [];
-
-        $this->assertEquals(
-            'yii.validation.boolean(value, messages, {"trueValue":true,"falseValue":false,"message":"attrB must be either \"true\" or \"false\".","skipOnEmpty":1,"strict":1});',
-            $validator->clientValidateAttribute($obj, 'attrB', new ViewStub())
-        );
-    }
-}
-
-class ViewStub extends View
-{
-    public function registerAssetBundle($name, $position = null)
-    {
     }
 }
